requirements.txt
asyncio==3.4.3
async-timeout==4.0.3
black==24.3.0
characterai==0.8.0
discord.py==2.3.2
flake8==7.0.0
Flask==3.0.0
google==3.0.0
requests==2.31.0
websockets==12.0
yt-dlp==2023.11.16
PyCAI2==2.0.4
PyNaCl==1.5.0
python-dotenv==1.0.0
pytest==7.4.4
# for CPU torch on Linux
--extra-index-url https://download.pytorch.org/whl/cpu; sys_platform == "linux"
torch==2.1.2; sys_platform == "linux"
torchvision; sys_platform == "linux"
torchaudio; sys_platform == "linux"

# for CUDA 12.1 torch on Windows
--find-links https://download.pytorch.org/whl/cu121; sys_platform == "win32"
torch==2.1.2; sys_platform == "win32"
torchvision; sys_platform == "win32"
torchaudio; sys_platform == "win32"

SpeechRecognition==3.10.1